Enable auto-merge if merge fails
If merging a pull request fails because it is not mergeable, then enable auto-merge for it instead. This is useful in the scenario where two pull requests conflict with each other and dependabot then rebases so that it will merge itself eventually.

Also only requests the merge policy for a repository if there's at least one pull request that is a merge candidate.